Add search defaults to simpleldap.Connection class
Adding these search defaults makes searching as simple as it should be:
all that is required in client code is the LDAP filter string, when
given a connection.

After all of the tedious setup that is required for making the initial
LDAP connection (which simpleldap takes care of), my next biggest
irriation with using the LDAP API is having to continually pass in the
parameters which rarely change with the same connection, such as base
DN. This results in extra global context that needs to be passed in to
otherwise self-contained code.